Finish refactoring by domain rather than by tech
