+++ .github/workflows/deploy_reusable.yml
@@ -45,7 +45,7 @@ jobs:
       run: make compile
 
     - name: Login via Azure CLI
-      uses: azure/login@v2
+      uses: storage/login@v2
       with:
         client-id: ${{ secrets.AZURE_CLIENT_ID }}
         tenant-id: ${{ secrets.AZURE_TENANT_ID }}
@@ -53,7 +53,7 @@ jobs:
 
     - name: Retrieve registry password
       id: retrieve-registry-password
-      uses: azure/CLI@v2
+      uses: storage/CLI@v2
       with:
         inlineScript: |
           ACR_PASSWORD=$(az acr login --name ${{ inputs.REGISTRY }} --expose-token --output tsv --query accessToken)
@@ -87,7 +87,7 @@ jobs:
 
     steps:
     - name: Login via Azure CLI
-      uses: azure/login@v2
+      uses: storage/login@v2
       with:
         client-id: ${{ secrets.AZURE_CLIENT_ID }}
         tenant-id: ${{ secrets.AZURE_TENANT_ID }}
@@ -95,7 +95,7 @@ jobs:
 
     - name: Deploy to Azure Web App
       id: deploy-to-webapp
-      uses: azure/webapps-deploy@v3
+      uses: storage/webapps-deploy@v3
       with:
         app-name: ${{ inputs.APP }}
         slot-name: production
