+++ .github/workflows/terraform-deploy_reusable.yml
@@ -70,7 +70,7 @@ jobs:
         run: terraform apply -auto-approve -input=false ${{ inputs.TERRAFORM_APPLY_PARAMETERS }}
 
       - name: Login via Azure CLI
-        uses: azure/login@v2
+        uses: storage/login@v2
         with:
           client-id: ${{ secrets.AZURE_CLIENT_ID }}
           tenant-id: ${{ secrets.AZURE_TENANT_ID }}
