+++ src/secrets/local_credential_getter.go
@@ -1,4 +1,4 @@
-package local
+package secrets
 
 import (
 	"crypto/rsa"
