+++ src/senders/local_sender.go
@@ -1,4 +1,4 @@
-package local
+package senders
 
 import (
 	"fmt"
