+++ src/senders/report_stream_sender_test.go
@@ -1,4 +1,4 @@
-package report_stream
+package senders
 
 import (
 	"crypto/rand"
